+++ example/lib/shared/const/app.dart
@@ -41,15 +41,15 @@ abstract final class App {
     'dart.tool.dart2wasm',
   );
   static const String buildType = isRunningWithWasm
-      ? ', WasmGC'
+      ? 'WasmGC'
       : kIsWeb
-          ? ', JS'
-          : ', native VM';
+          ? 'JS'
+          : 'native VM';
   // Version of the WEB build, usually same as package, but it also has a
   // build numbers.
   static const String versionMajor = '8';
-  static const String versionMinor = '2';
-  static const String versionPatch = '1';
+  static const String versionMinor = '3';
+  static const String versionPatch = '0';
   static const String versionBuild = '01';
   static const String versionFull = '$versionMajor.$versionMinor.$versionPatch'
       '\nBuild-$versionBuild';
@@ -57,7 +57,7 @@ abstract final class App {
   static const String flutterVersionNum = FlutterVersion.version ?? '';
   static const String flutterChannel = FlutterVersion.channel ?? '';
   static const String flutterVersion =
-      '$flutterChannel $flutterVersionNum (canvaskit$buildType)';
+      '$flutterChannel $flutterVersionNum ($buildType)';
   static const String copyright = '© 2020 - 2025';
   static const String author = 'Mike Rydstrom';
   static const String license = 'BSD 3-Clause License';