+++ example/lib/shared/widgets/universal/showcase_material.dart
@@ -1729,7 +1729,7 @@ class _DropdownButtonFormFieldShowcaseState
               ),
             ),
           DropdownButtonFormField<String>(
-            value: selectedItem,
+            initialValue: selectedItem,
             onChanged: (String? value) {
               setState(() {
                 selectedItem = value ?? '1 DropdownButtonFormField';
@@ -1812,48 +1812,28 @@ class _DropDownMenuShowcaseState extends State<DropDownMenuShowcase> {
             },
             dropdownMenuEntries: const <DropdownMenuEntry<IconData>>[
               DropdownMenuEntry<IconData>(
-                style: ButtonStyle(
-                  padding: WidgetStatePropertyAll<EdgeInsetsGeometry>(
-                      EdgeInsets.symmetric(horizontal: 12)),
-                ),
                 label: 'Alarm settings',
                 leadingIcon: Icon(Icons.alarm),
                 value: Icons.alarm,
               ),
               DropdownMenuEntry<IconData>(
-                style: ButtonStyle(
-                  padding: WidgetStatePropertyAll<EdgeInsetsGeometry>(
-                      EdgeInsets.symmetric(horizontal: 12)),
-                ),
                 label: 'Disabled settings',
                 leadingIcon: Icon(Icons.settings),
                 enabled: false,
                 value: Icons.settings,
               ),
               DropdownMenuEntry<IconData>(
-                style: ButtonStyle(
-                  padding: WidgetStatePropertyAll<EdgeInsetsGeometry>(
-                      EdgeInsets.symmetric(horizontal: 12)),
-                ),
                 label: 'Cabin overview',
                 leadingIcon: Icon(Icons.cabin),
                 value: Icons.cabin,
               ),
               DropdownMenuEntry<IconData>(
-                style: ButtonStyle(
-                  padding: WidgetStatePropertyAll<EdgeInsetsGeometry>(
-                      EdgeInsets.symmetric(horizontal: 12)),
-                ),
                 label: 'Surveillance view',
                 leadingIcon: Icon(Icons.camera_outdoor_rounded),
                 // value: 'four',
                 value: Icons.camera_outdoor_rounded,
               ),
               DropdownMenuEntry<IconData>(
-                style: ButtonStyle(
-                  padding: WidgetStatePropertyAll<EdgeInsetsGeometry>(
-                      EdgeInsets.symmetric(horizontal: 12)),
-                ),
                 label: 'Water alert',
                 leadingIcon: Icon(Icons.water_damage),
                 // value: 'five',
@@ -2676,8 +2656,8 @@ class _TextFieldShowcaseState extends State<TextFieldShowcase> {
 
   @override
   Widget build(BuildContext context) {
-    final bool isDefaultDecoration =
-        Theme.of(context).inputDecorationTheme == const InputDecorationTheme();
+    final bool isDefaultDecoration = Theme.of(context).inputDecorationTheme ==
+        const InputDecorationThemeData();
 
     return RepaintBoundary(
       child: Column(