+++ CHANGELOG.md
@@ -2,6 +2,51 @@
 
 All changes to the **FlexColorScheme** (FCS) package are documented here.
 
+## 8.3.0
+
+**August 15, 2025**
+
+This release of FlexColorScheme adds support for breaking changes in Flutter v3.35.0 and also requires at least this version.
+
+### Package
+
+**FIX**
+
+* Fix the in Flutter v3.35 broken API for `BottomAppBarTheme`, see [issue #283](https://github.com/rydmike/flex_color_scheme/issues/283)
+* Fix the Flutter v3.35 broken API for `AppBarTheme`, migrate to use new `AppBarThemeData`. For more information, see the discussion in [#PR285](https://github.com/rydmike/flex_color_scheme/pull/285)
+* Fix the Flutter v3.35 broken API for `InputDecorationTheme`, migrate to use new `InputDecorationThemeData`. For more information, see the discussion in [#PR285](https://github.com/rydmike/flex_color_scheme/pull/285)
+* Added a temporary workaround for the `DropdownMenu`'s width issue [#170970](https://github.com/flutter/flutter/issues/170970) so that the correct default width style for the `DropdownMenu` is retained. This issue affects Flutter 3.32.0 to at least Flutter 3.35.1. There is a fix PR in the master channel [#PR169438](https://github.com/flutter/flutter/pull/169438), but it has not landed in Flutter 3.35.1. See also FCS issue report [#286](https://github.com/rydmike/flex_color_scheme/issues/286). 
+
+**CHORE**
+
+* Updated tests to use the new `BottomAppBarThemeData` API.
+* Updated tests to use the new `AppBarThemeData` API.
+* Updated tests to use the new `InputDecorationThemeData` API.
+* The test for `FlexSubThemeData.sliderShowValueIndicator` was changed from in **Flutter 3.35** deprecated value `ShowValueIndicator.always` to `ShowValueIndicator.onDrag`, which is the new equivalent value in Flutter 3.35.
+* Bump packages to latest versions.
+
+**FEATURE**
+
+* **DropdownMenu**: Added theming properties for the `DropdownMenu`'s maximum size.
+  * `FlexSubThemesData` got the new property `dropdownMenuMaximumSize`.
+  * `FlexSubThemes.dropdownMenuTheme` got the new property `maximumSize`.
+  * Both of type `WidgetStateProperty<Size?>?`
+  * This was added to allow overriding the fix for the `DropdownMenu` width issue [#170970](https://github.com/flutter/flutter/issues/170970).
+  * The `FlexSubThemes.dropdownMenuTheme` property `maximumSize` temporarily default fallbacks to `const WidgetStatePropertyAll<Size>(Size.infinite)` if undefined. This results in the same default and correct style that Flutter had in 3.29.x and earlier versions. The default fallback will be removed in a future version of **FlexColorScheme** when the issue is fixed in Flutter SDK stable release.
+  * This property is not available in the **Themes Playground** app, you can only override it via FlexColorScheme APIs.
+
+
+### Playground
+
+**CHORE**
+* Change deprecated `Switch.activeColor` to `Switch.activeThumbColor` in `ColorNameValue`.
+* Use the new `RadioGroup` API in the Playground Theme Simulator in the Material-3 example app.
+* Bump packages to latest versions. 
+
+* **FEATURE**
+* The slider indicator value selection got the new selections `ShowValueIndicator.onDrag` and `ShowValueIndicator.alwaysShow`, which are new values in Flutter 3.35. The old value `ShowValueIndicator.always` is deprecated in Flutter 3.35 and will be removed in a future version of Flutter SDK and the Playground app.
+
+ 
 ## 8.2.0
 
 **March 23, 2025**